valuai-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
pgvector==0.3.5
alembic==1.13.3
groq==0.11.0
google-generativeai==0.8.3
firecrawl-py==1.5.0
httpx==0.27.2
openpyxl==3.1.5
python-multipart==0.0.12
python-dotenv==1.0.1
tenacity==8.5.0
structlog==24.4.0
fpdf2==2.7.9
jinja2==3.1.4
pillow==10.4.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
